refactor: 替换魔术数字为常量并改进类型安全
- 将 constants.rs 中的 SC01/SC08/SC16 常量类型从 usize 改为 i32
- 在通道处理逻辑中使用 N_SBF 和 N_DWRD_SBF 常量替代魔术数字
- 添加返回值状态注释 (-1=无效, 0=不可见, 1=可见)
- 引入 R2D 常量替代硬编码的弧度转角度值
- 使用 MAX_SAT 和 MAX_CHAN 常量优化卫星通道分配逻辑
- 更新星历数组维度声明为 EPHEM_ARRAY_SIZE

+++ src/constants.rs
@@ -61,8 +61,8 @@ pub const CODE_FREQ: f64 = 1.023e6;
 pub const CARR_TO_CODE: f64 = 1.0 / 1540.0;
 
 // Sampling data format
-pub const SC01: usize = 1;
-pub const SC08: usize = 8;
-pub const SC16: usize = 16;
+pub const SC01: i32 = 1;
+pub const SC08: i32 = 8;
+pub const SC16: i32 = 16;
 
 pub const EPHEM_ARRAY_SIZE: usize = 15; // for daily GPS broadcast ephemers file (brdc)
